Fix time diff

+++ pepperl_fuchs_r2000/src/rosnode/r2000_node.h
@@ -82,7 +82,7 @@ private:
     //! Pointer to driver
     R2000Driver* driver_;
 
-    uint64_t base_time_;
+    double base_time_;
     ros::Time ros_base_time_;
     bool first_data_;
 };
